Re-created everything
Now it is an ASP NET Core web API project
and an Angular frontend.

Had to make the web API first with the
"dotnet new angular" command then I nuked
everything in the "ClientApp" and made
a new Angular app within it.

Had to modify the Startup.cs.

Refer to "how-to-run.txt" for what is needed
to run both.